ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain==1.2.0
dashscope==1.20.11
paddlepaddle==3.0.0
paddleocr
PyPDF2==3.0.1
python-docx==1.1.2
pydantic==2.8.2
httpx==0.27.0
sse-starlette==2.1.3
python-multipart==0.0.9
python-dotenv==1.0.1
langchain-community
opencv-python==4.10.0.84
pymysql==1.1.1
cryptography==43.0.0
redis==5.2.1
DBUtils==3.1.0
chromadb==0.5.5